requirements.txt
crewai>=0.70.0
litellm>=1.30.0
python-dotenv>=1.0.0
reportlab>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
pandas>=2.1.0
numpy>=1.24.0
textblob>=0.17.1
vaderSentiment>=3.3.2
youtube-transcript-api>=0.6.1
pytube>=15.0.0
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
sentence-transformers>=2.2.0
Pillow>=10.0.0
opencv-python>=4.8.0
transformers>=4.35.0
torch>=2.1.0
